Restructure components into corresponding folders and remove unused imports
