+++ src/app/auth/confirm/route.ts
@@ -1,3 +1,4 @@
+// delete if auth is not used
 import { type EmailOtpType } from "@supabase/supabase-js";
 import { type NextRequest, NextResponse } from "next/server";
