--- src/app/order-confirmation/page.tsx
@@ -1,3 +1,0 @@
-export default function page() {
-  return <div>confirmation page</div>;
-}
