--- src/components/Item/Item.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Item() {
-  return <div>This is an individual item</div>;
-}
