+++ src/components/main-content/CareersPreview/CareersPreview.tsx
@@ -1,5 +1,11 @@
-import { CareerPreviewProps } from "../../types";
-import { Button, Card, CardBody, CardFooter, Typography } from "../../tailwind";
+import { CareerPreviewProps } from "../../../types";
+import {
+  Button,
+  Card,
+  CardBody,
+  CardFooter,
+  Typography,
+} from "../../../tailwind";
 import React from "react";
 
 interface CareersProps {
