+++ src/components/main-content/HeroImage/HeroImage.tsx
